Created consolidated index.js for all services hooks
