--- src/common/alter-data-split.js
@@ -1,14 +1,0 @@
-
-import _pluck from './_pluck';
-import _remove from './_remove';
-import _setFields from './_set-fields';
-import _transformItems from './_transform-items';
-import _traverse from './_traverse';
-
-export default {
-  _pluck,
-  _remove,
-  _setFields,
-  _transformItems,
-  _traverse
-};
