--- src/services/alter-data-split.js
@@ -1,16 +1,0 @@
-
-import lowerCase from './lower-case';
-import pluck from './pluck';
-import remove from './remove';
-import setCreatedAt from './set-created-at';
-import setUpdatedAt from './set-updated-at';
-import traverse from './traverse';
-
-export default {
-  lowerCase,
-  pluck,
-  remove,
-  setCreatedAt,
-  setUpdatedAt,
-  traverse
-};
