--- src/services/query-params-split.js
@@ -1,9 +1,0 @@
-// temporary
-
-import removeQuery from './remove-query';
-import pluckQuery from './pluck-query';
-
-export default {
-  removeQuery,
-  pluckQuery
-};
