--- src/services/utilities-split.js
@@ -1,17 +1,0 @@
-// temporary
-
-import callbackToPromise from './callback-to-promise';
-import client from './client';
-import debug from './debug';
-import disable from './disable';
-import promiseToCallback from './promise-to-callback';
-import setSlug from './set-slug';
-
-export default {
-  callbackToPromise,
-  client,
-  debug,
-  disable,
-  promiseToCallback,
-  setSlug
-};
