--- src/services/utils.js
@@ -1,16 +1,0 @@
-
-import checkContext from './check-context';
-import checkContextIf from './check-context-if';
-import getByDot from '../common/get-by-dot';
-import getItems from './get-items';
-import replaceItems from './replace-items';
-import setByDot from '../common/set-by-dot';
-
-export default {
-  checkContext,
-  checkContextIf,
-  getByDot,
-  getItems,
-  replaceItems,
-  setByDot
-};
