--- src/services/validate-split.js
@@ -1,9 +1,0 @@
-// temporary
-
-import validate from './validate';
-import validateSchema from './validate-schema';
-
-export default {
-  validate,
-  validateSchema
-};
